#31 Separate request and response classes
Moves requests and response classes into dedicated modules.
