+++ docs/examples/direct/string_response.py
@@ -1,4 +1,4 @@
-from pygrestqlambda.aws.lambda_function.rest_api_gateway_proxy_integration import Response
+from pygrestqlambda.aws.lambda_function.rest_api_gateway_proxy_integration.response import Response
 
 response = Response(
     headers={'Content-Type': 'text/html'},
