+++ tests/integration/utils/lambda_function_docker/app.py
@@ -3,7 +3,7 @@ Lambda function handler to be used in integration tests
 """
 
 import logging
-from pygrestqlambda.aws.lambda_function.rest_api_gateway_proxy_integration import Response
+from pygrestqlambda.aws.lambda_function.rest_api_gateway_proxy_integration.response import Response
 
 def handler(event, context):
     """
